+++ scour.js
@@ -66,11 +66,30 @@ scour.prototype = {
    * Chaining methods:
    * (Section) These methods are used to traverse nested structures. All these
    * methods return [scour] instances, making them suitable for chaining.
+   *
+   * #### On null values
+   * Note that `undefined`, `false` and `null` values are still [scour]-wrapped
+   * when returned from [go()], [at()] and [find()].
+   *
+   *     list = [ { name: 'Homer' }, { name: 'Bart' } ]
+   *
+   *     scour(list).at(4)         // => [ scour undefined ]
+   *     scour(list).at(4).value   // => undefined
+   *
+   * This is done so that you can chain methods safely even when something is null.
+   * This behavior is consistent with what you'd expect with jQuery.
+   *
+   *     data = { users: { ... } }
+   *     db = scour(data)
+   *
+   *     db.go('blogposts').map((post) => post.get('title'))
+   *     // => []
    */
 
   /**
    * go : go(keypath...)
    * Navigates down to a given `keypath`. Always returns a [scour] instance.
+   * Rules [on null values] apply.
    *
    *     data =
    *       { users:
@@ -103,7 +122,7 @@ scour.prototype = {
   go () {
     const keypath = normalizeKeypath(arguments, true)
     const result = this.get.apply(this, keypath)
-    return this._get(result, keypath, true)
+    return this._get(result, keypath)
   },
 
   /**
@@ -122,7 +141,8 @@ scour.prototype = {
 
   /**
    * Returns the item at `index`. This differs from `go` as this searches by
-   * index, not by key. This returns a the raw value, unlike [getAt()].
+   * index, not by key. This returns a the raw value, unlike [getAt()]. Rules
+   * [on null values] apply.
    *
    *     users =
    *       { 12: { name: 'steve' },
@@ -246,7 +266,7 @@ scour.prototype = {
    * Returns the first value that matches `conditions`.  Supports MongoDB-style
    * queries. For reference, see [MongoDB Query Operators][query-ops]. Also
    * see [filter()], as this is functionally-equivalent to the first result of
-   * `filter()`.
+   * `filter()`. Rules [on null values] apply.
    *
    * [query-ops]: https://docs.mongodb.org/manual/reference/operator/query/
    *
@@ -276,7 +296,7 @@ scour.prototype = {
 
   last () {
     var len = this.len()
-    if (len > 0) return this.at(len - 1)
+    return this.at(len - 1)
   },
 
   /**
@@ -753,7 +773,6 @@ scour.prototype = {
    */
 
   _get (result, keypath) {
-    if (typeof result === 'undefined' || result === null) return result
     return this.replace(result, {
       keypath: this.keypath.concat(keypath)
     })
@@ -772,7 +791,7 @@ scour.prototype = {
 
   replace (value, options) {
     const op = options || {}
-    return new scour(value || this.value, {
+    return new scour(value, {
       root:
         typeof op.root !== 'undefined' ? op.root : this.root,
       keypath: