+++ test/todo_test.js
@@ -20,8 +20,8 @@ describe('todo', function () {
 
   describe('array', function () {
     it('.at', done)
-    it('.go and .at returning objects, ALWAYS')
-    it('.getAt')
+    it('.go and .at returning objects, ALWAYS', done)
+    it('.getAt', done)
     it('.gather')
     it('.slice')
     it('.initial')